Identipy refimpl bug fixes + optimizations
There were two major bugs in IdentiPy that were copied over in the reference implementation.
- Incorrect secondary mask at the following line:
nacc = numpy.where(dist[mask1] / mz_array[ind[mask1]] * 1e6 > acc)[0]
Which leads to:
mask2[nacc] = False # note: mask1 and mask2 are the same thing...

This is incorrect because 'where' will return indices that only corresponds to the 'true' entries in mask1,
but mask1 (aka mask2) contains both true and false entries (it's a binary mask). Hence, a setup like
dist = [1, 2, 3]
mask1 = [false, true, false]
mz_array = [10, 20, 30]
and acc such that numpy.where is true everywhere, will yield:
numpy.where([2] / [20] * 1e6 > acc)[0] = 0
Which results in
mask2[0] = false
for a final mask of [false, true, false] when it was supposed to be [false, false, false].

This was the reason why the crnhs was not able to yield the same results as the reference implementation, along with the reference implementation
using a theoretical spectrum generator providing peaks in decreasing order (due to the above bug, the order of peaks yields different final masks and thus scores).

- Incorrect cutoff code when using ppm cutoff
Since the distance_upper_bound for scipy's cKDTree must be the same value across the board, the developers elected to use a default value of
float(acc) / 1e6 * 2000 for the cutoff provided by the tree, and to then further remove those elements which passed that test but don't pass the true ppm cutoff test.
This effectively means assuming a peak mass of 2000 for the upper bound, which will be too stingent for peaks that have a mass beyond 2000 m/z and result in false negatives.
This significantly affects the final results in practice and notably reduces the TDA-FDR score compared to the corrected version.